fix():控制台修改 & 结构整理
