+++ atom-service/atom-service-operator/atom-service-operator-consumers/src/main/java/com/lamp/atom/service/operator/consumers/controller/ModelController.java
@@ -18,6 +18,7 @@ import com.lamp.atom.service.operator.service.ModelService;
 import com.lamp.atom.service.operator.consumers.utils.ResultObjectEnums;
 import com.lamp.decoration.core.result.ResultObject;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.dubbo.config.annotation.Reference;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -33,8 +34,7 @@ import java.util.Objects;
 @RestController("modelController")
 public class ModelController {
 
-    @Autowired
-    @Qualifier("modelService")
+    @Reference
     private ModelService modelService;
 
     /**
