+++ atom-service/atom-service-operator/atom-service-operator-consumers/src/main/java/com/lamp/atom/service/operator/consumers/controller/OperatorEventController.java
@@ -16,6 +16,7 @@ import com.lamp.atom.service.domain.OperatorStatus;
 import com.lamp.atom.service.operator.entity.OperatorEntity;
 import com.lamp.atom.service.operator.service.OperatorService;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.dubbo.config.annotation.Reference;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -28,8 +29,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController("operatorEventController")
 public class OperatorEventController {
 
-    @Autowired
-    @Qualifier("operatorService")
+    @Reference
     private OperatorService operatorService;
 
     /**
