+++ atom-service/atom-service-operator/atom-service-operator-entity/src/main/java/com/lamp/atom/service/operator/entity/AvaliablePortEntity.java
@@ -1,0 +1,11 @@
+package com.lamp.atom.service.operator.entity;
+
+import lombok.Data;
+
+@Data
+public class AvaliablePortEntity extends BaseEntity {
+
+    private static final long serialVersionUID = 6178707504251128831L;
+
+    private Integer port;
+}
