+++ atom-service/atom-service-operator/atom-service-operator-entity/src/main/java/com/lamp/atom/service/operator/entity/SourceAccountEntity.java
@@ -1,0 +1,4 @@
+package com.lamp.atom.service.operator.entity;
+
+public class SourceAccountEntity {
+}
