+++ atom-service/atom-service-operator/atom-service-operator-provider/src/main/java/com/lamp/atom/service/operator/provider/service/impl/OperatorServiceImpl.java
@@ -14,14 +14,14 @@ package com.lamp.atom.service.operator.provider.service.impl;
 import com.lamp.atom.service.operator.entity.OperatorEntity;
 import com.lamp.atom.service.operator.provider.mapper.OperatorMapper;
 import com.lamp.atom.service.operator.service.OperatorService;
+import org.apache.dubbo.config.annotation.Service;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
+@Service
 @Transactional
-@Service("operatorService")
 public class OperatorServiceImpl implements OperatorService {
 
     @Autowired
