+++ atom-service/atom-service-operator/atom-service-operator-service/src/main/java/com/lamp/atom/service/operator/service/AvaliablePortService.java
@@ -1,0 +1,36 @@
+/*
+ *Copyright (c) [Year] [name of copyright holder]
+ *[Software Name] is licensed under Mulan PubL v2.
+ *You can use this software according to the terms and conditions of the Mulan PubL v2.
+ *You may obtain a copy of Mulan PubL v2 at:
+ *         http://license.coscl.org.cn/MulanPubL-2.0
+ *THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
+ *EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
+ *MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
+ *See the Mulan PubL v2 for more details.
+ */
+package com.lamp.atom.service.operator.service;
+
+import com.lamp.atom.service.operator.entity.AvaliablePortEntity;
+
+public interface AvaliablePortService {
+    /**
+     * 添加可用端口
+     * @param avaliablePortEntity
+     */
+    Integer insertAvaliablePortEntity(AvaliablePortEntity avaliablePortEntity);
+
+    /**
+     * 修改可用端口
+     * @param avaliablePortEntity
+     * @return
+     */
+    Integer updateAvaliablePortEntity(AvaliablePortEntity avaliablePortEntity);
+
+    /**
+     * 查询单个可用端口
+     * @param avaliablePortEntity
+     * @return
+     */
+    AvaliablePortEntity queryAvaliablePortEntity(AvaliablePortEntity avaliablePortEntity);
+}
