+++ atom-service/atom-service-space/atom-service-space-consumers/src/main/java/com/lamp/atom/service/space/consumer/utils/ResultObjectEnums.java
