+++ atom-service/atom-service-space/atom-service-space-entity/src/main/java/com/lamp/atom/space/operator/entity/BaseEntity.java
