+++ atom-service/atom-service-space/atom-service-space-service/src/main/java/com/lamp/atom/space/operator/service/SpaceService.java
@@ -1,0 +1,4 @@
+package com.lamp.atom.space.operator.service;
+
+public interface SpaceService {
+}
