--- atom-service/atom-service-task/src/main/java/com/lamp/atom/service/task/entity/BaseEntity.java
@@ -1,46 +1,0 @@
-/*
- *Copyright (c) [Year] [name of copyright holder]
- *[Software Name] is licensed under Mulan PubL v2.
- *You can use this software according to the terms and conditions of the Mulan PubL v2.
- *You may obtain a copy of Mulan PubL v2 at:
- *         http://license.coscl.org.cn/MulanPubL-2.0
- *THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
- *EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
- *MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
- *See the Mulan PubL v2 for more details.
- */
-package com.lamp.atom.service.task.entity;
-
-import lombok.Data;
-
-import java.io.Serializable;
-import java.util.Date;
-
-@Data
-public class BaseEntity implements Serializable {
-
-    /**
-     * 序列号
-     */
-    private static final long serialVersionUID = -7162490774296452711L;
-
-    /**
-     * 主键id
-     */
-    private Long id;
-
-    /**
-     * 创建时间
-     */
-    private Date createTime;
-
-    /**
-     * 修改时间
-     */
-    private Date updateTime;
-
-    /**
-     * 状态：0正常1删除
-     */
-    private Integer deleteFlag;
-}
