--- atom-space/atom-space-consumers/src/main/java/com/lamp/atom/space/operator/AtomSpaceOperatorConsumerApplication.java
@@ -1,31 +1,0 @@
-/*
- *Copyright (c) [Year] [name of copyright holder]
- *[Software Name] is licensed under Mulan PubL v2.
- *You can use this software according to the terms and conditions of the Mulan PubL v2.
- *You may obtain a copy of Mulan PubL v2 at:
- *         http://license.coscl.org.cn/MulanPubL-2.0
- *THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
- *EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
- *MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
- *See the Mulan PubL v2 for more details.
- */
-package com.lamp.atom.space.operator;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-@Slf4j
-@SpringBootApplication
-@EnableTransactionManagement
-public class AtomSpaceOperatorConsumerApplication {
-    public static void main(String[] args) {
-        try {
-            SpringApplication.run(AtomSpaceOperatorConsumerApplication.class, args);
-            log.info("{} 启动成功", AtomSpaceOperatorConsumerApplication.class.getSimpleName());
-        } catch (Exception e) {
-            log.error(e.getMessage() , e);
-        }
-    }
-}
